phase 1
Đã có xác thực token, bảo vệ API và WebSocket.
Đăng ký, đăng nhập, đăng xuất, quản lý phiên tốt.
Giao diện realtime, lưu lịch sử chat, bảo mật cơ bản ổn.
