+++ controllers/chat_controller.py
@@ -4,29 +4,32 @@ from db.models import User, Message
 from db.database import get_db, SessionLocal
 from authentication.auth import get_current_user
 from fastapi.templating import Jinja2Templates
-from main import ConnectionManager
 from sqlalchemy.orm import Session
 from fastapi.responses import HTMLResponse
 from fastapi.exceptions import HTTPException
 import json
+from utils.connection_manager import ConnectionManager
+connection_manager = ConnectionManager()
+
+router = APIRouter()
 
 templates = Jinja2Templates(directory="templates")
 router = APIRouter()
 
-@app.websocket("/message")
+@router.websocket("/message")
 async def websocket_endpoint(websocket: WebSocket):
+    await connection_manager.connect(websocket)
     token = websocket.cookies.get("token")
     db = SessionLocal()
     user = db.query(User).filter(User.token == token).first()
     if not user:
         await websocket.close()
         return
-    await ConnectionManager.connect(websocket)
     db = SessionLocal()
     # Gửi lại các tin nhắn cũ cho client mới
     old_messages = db.query(Message).order_by(Message.created_at).all()
     for msg in old_messages:
-        await ConnectionManager.send_message(
+        await connection_manager.send_message(
             websocket,
             json.dumps({"isMe": False, "data": msg.content, "username": msg.username})
         )
@@ -38,7 +41,7 @@ async def websocket_endpoint(websocket: WebSocket):
             message = decoded.get("message")
             # Kiểm tra username hợp lệ
             if not username:
-                await ConnectionManager.send_message(
+                await connection_manager.send_message(
                     websocket,
                     json.dumps({"isMe": True, "data": "Username is missing!", "username": "system"})
                 )
@@ -47,19 +50,19 @@ async def websocket_endpoint(websocket: WebSocket):
             msg = Message(username=username, content=message)
             db.add(msg)
             db.commit()
-            await ConnectionManager.broadcast(websocket, data)
+            await connection_manager.broadcast(websocket, data)
     except WebSocketDisconnect:
-        id = ConnectionManager.disconnect(websocket)
+        id = connection_manager.disconnect(websocket)
         db.close()
         return RedirectResponse("/")
     
-@app.get("/join", response_class=HTMLResponse)
+@router.get("/join", response_class=HTMLResponse)
 def join_form(request: Request, user: User = Depends(get_current_user)):
     username = user.username
     response = templates.TemplateResponse("index.html", {"request": request, "username": username})
     return response
 
-@app.post("/join")
+@router.post("/join")
 async def join(request: Request):
     form = await request.form()
     username = form["username"]
@@ -73,13 +76,13 @@ async def join(request: Request):
         response.set_cookie("token", user.token)
     return response
 
-@app.get("/chat", response_class=HTMLResponse)
+@router.get("/chat", response_class=HTMLResponse)
 def chat_room(request: Request, user: User = Depends(get_current_user)):
     username = user.username
     response = templates.TemplateResponse("index.html", {"request": request, "username": username})
     return response
         
-@app.post("/api/message")
+@router.post("/api/message")
 def create_message(
     data: dict = Body(...),
     token: str = Cookie(None),
