+++ templates/index.html
@@ -56,7 +56,7 @@
     function scrollMessagesToBottom() {
       const messages = document.getElementById('messages');
       if (messages.lastElementChild) {
-        messages.lastElementChild.scrollIntoView({ behavior: 'smooth' });
+        messages.lastElementChild.scrollIntoView();
       }
     }
